fix(nevermore-test-runner): validate jest's counts, and stop double-counting suites
Two ways this could still report a pass on a failing run, and the reason it
kept happening.

findAggregatedResult accepted any table with a numeric numTotalTests, and every
other count defaulted to 0 when it could not be read. So a jest-lua rename of
numFailedTests, numFailedTestSuites, numRuntimeErrorTestSuites or
wasInterrupted reproduced the original bug exactly: all counts zero, verdict
pass, green CI. Not hypothetical — the port already spells the per-suite field
numFailingTests next to the aggregate numFailedTests. Validating field names
one at a time only ever covers the names known when the check was written, so
the sum is asserted instead: total == passed + failed + skipped holds by
construction (helpers.lua:91 derives numTotalTests from exactly those
per-suite counts), and it catches a rename of any count, named or not. Every
count the verdict reads must now also be present and numeric, and a mismatch
fails closed.

suitesFailed added numFailedTestSuites to numRuntimeErrorTestSuites, but
jest-lua increments both for one suite that failed to run — the runtime-error
counter unconditionally, then numFailedTestSuites through the separate
`numFailingTests > 0 or testExecError` branch. One broken suite counted twice,
and a single-spec package reported "2 of 1 suite(s) failed". It reports
numFailedTestSuites now and keeps numRuntimeErrorTestSuites as its own verdict
term, which is the only count that sees a suite both skipped and broken.

Both bugs, and the two before them, were untestable rather than subtle: a
module-scope require("Jest") stops the whole file loading in any harness
without Jest and the Nevermore loader. The logic moved to NevermoreTestResults
— no loader, no Jest, no print or warn, plain tables in and out — and
NevermoreTestRunnerUtils keeps the Roblox-facing half and re-exports the types,
so callers see no change. A dependency-free Luau module needs no loader line;
312 of the repo's package modules already omit it.

The tests run under Lune, from the package: npm test. 44 assertions, no harness
and no mocks, covering both directions of every verdict term and the shapes
that must fail closed. Their proper home is a jest test target, which needs a
placeId this cannot mint; Lune needs no credentials, so Lune is what exists.

Also fixes a red CI job: a doc comment on a local function used --[=[ … ]=],
which moonwave-extractor treats as a doc block and rejects with "Function
requires @within tag" — and it aborts on the first diagnostic, so nothing else
was being checked either. Local helpers use --[[ … ]] now, and *.test.luau is
excluded from lint:luau since @lune/* requires cannot resolve under luau-lsp.
